+++ src/features/landing/pricing/types/index.ts
@@ -1,7 +1,7 @@
 import type {
   PassSingleProduct,
   CreditPackProduct,
-} from "@/shared/pricing/types";
+} from "@/shared/landing/pricing/types";
 
 export type Plan = PassSingleProduct;
 export type TokenAdd = CreditPackProduct;
