+++ src/features/landing/pricing/ui/PlanCardList.tsx
@@ -4,7 +4,7 @@ import { useSuspenseQuery } from "@tanstack/react-query";
 
 import { Plan } from "@/features/landing/pricing/types";
 import { useEducatorCheckoutNavigation } from "@/features/landing/pricing/hooks/useEducatorCheckoutNavigation";
-import { pricingQueries } from "@/shared/pricing/api/query";
+import { pricingQueries } from "@/shared/landing/pricing/api/query";
 import { useCheckoutStore } from "@/shared/common/store/useCheckoutStore";
 
 import { PlanCard } from "./PlanCard";
